+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `CoreVideo` / `objc2-core-video`.
   - `EventKitUI` / `objc2-event-kit-ui`.
   - `IOSurface` / `objc2-io-surface`.
+  - `MetalPerformanceShaders` / `objc2-metal-performance-shaders`.
   - `ScreenSaver` / `objc2-screen-saver`.
 * Added `#[must_use]` attributes where the C headers have them.